[TOW-41] 레벨 선택 화면 리팩터링
* CardView 분리
* Store 주입 대신 isSelected 값으로 변경
